+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Update RevitSystemTestBase to make it more user friendly
 * Add two new nodes to ImportInstance because of RevitAPI changes
 * Make Nodes FaceAnalysisDisplay.ByViewAndFaceAnalysisData, PointAnalysisDisplay.ByViewAndPointAnalysisData, VectorAnalysisDisplay.ByViewAndVectorAnalysisData invisible in Dynamo because SurfaceData, PointData and VectorData was set obsolete in DynamoCore.
+* Improve RevitSystemTestBase that DynamoRevit custom node developers could preload their nodes for test in Code.
 
 ## 0.1.0
 * The version of master branch will be start with 0.1.x, and Revit2020 branch will still be 0.0.xx.